requirements.txt
# Core framework
streamlit
requests

# Data handling
pandas
numpy

# Visualization
matplotlib
seaborn

# Machine Learning
scikit-learn

# Report generation
reportlab

# Environment variables
python-dotenv

# LangChain & Groq
langchain
langchain-groq
langgraph
langchain-experimental

# Hugging Face Hub (optional, but good to have)
huggingface-hub

# Extra (to avoid common issues)
openpyxl   # for Excel file support
pyyaml     # sometimes needed by langchain
